Int casting

+++ DCMotor.py
@@ -59,4 +59,4 @@ class DCMotor:
 
     @staticmethod
     def _map_value(x, in_min, in_max, out_min, out_max):
-        return (x - in_min) * (out_max - out_min) / (in_max - in_min) + out_min
+        return int((x - in_min) * (out_max - out_min) / (in_max - in_min) + out_min)
